refactor: 보충강의 듣고 폴더 구조 정리 - src 바로밑의 components폴더에는 공용 컴포넌트 - 페이지에서 쓰는 components는 각 페이지 폴더아래에서 보관
