+++ src/pages/Items/components/ProductThumbnail/index.jsx
