+++ src/pages/Items/components/ProductThumbnail/styles.module.scss
