+++ src/pages/Items/detail/index.jsx
@@ -1,5 +1,5 @@
 import Temporary from "@components/Temp";
 
-export default function detail() {
+export default function ItemDetail() {
   return <Temporary title="아이템 상세페이지" />;
 }
