+++ src/pages/auth/login/index.jsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import { useAuth } from "@context/useAuth";
 import useForm from "@hooks/useForm";
-import AuthContainer from "@components/Auth";
+import AuthContainer from "../components/AuthContainer";
 import Input from "@components/Input";
 import Button from "@components/Button";
 import { VALIDATION_MESSAGES, VALIDATION_REGEX } from "@util/validation";
